feat: add sort posts button
