+++ apps/front/src/pages/MainPage.tsx
@@ -1,7 +1,7 @@
 import styled from "@emotion/styled";
 import { Contribution, Introduce, PopularPosts, Projects } from "@/widgets";
 
-export function Main() {
+export function MainPage() {
   return (
     <Container>
       <Introduce />
