+++ apps/front/src/pages/StudyPage.tsx
@@ -1,4 +1,4 @@
-export function Study() {
+export function StudyPage() {
   return <div>Study 화면 개발 중...</div>;
 }
