+++ apps/front/src/pages/VisitorPage.tsx
@@ -1,4 +1,3 @@
-export function Visitor() {
+export function VisitorPage() {
   return <div>Visitor 화면 개발 중...</div>;
 }
-
