+++ apps/front/src/widgets/index.ts
@@ -4,3 +4,4 @@ export { PopularPosts } from "./PopularPosts";
 export { Projects } from "./Projects";
 export { Contribution } from "./Contribution";
 export { Footer } from "./Footer";
+export { AllPosts } from "./AllPosts";
